pyproject.toml
[project]
name = "mcp-timesheet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.4",
    "httpx>=0.28.1",
    "jira>=3.10.5",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.7",
    "mcp>=1.26.0",
    "openai>=2.16.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]
